@@ -5,7 +5,7 @@ homepage: https://github.com/Tkko/Flutter_dismissible_page
 repository: https://github.com/Tkko/Flutter_dismissible_page
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
@@ -16,5 +16,5 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  pedantic: ^1.9.0
+  pedantic: ^1.11.0
 
