@@ -5,7 +5,7 @@ class TransparentRoute<T> extends PageRoute<T>
   final WidgetBuilder builder;
 
   @override
-  final String title;
+  final String? title;
 
   /// Builds the primary contents of the route.
   @override
@@ -14,10 +14,10 @@ class TransparentRoute<T> extends PageRoute<T>
   final Color backgroundColor;
 
   TransparentRoute({
-    @required this.builder,
-    @required this.backgroundColor,
+    required this.builder,
+    required this.backgroundColor,
     this.title,
-    RouteSettings settings,
+    RouteSettings? settings,
     this.maintainState = true,
     bool fullscreenDialog = true,
   })  : assert(builder != null),
